refactor profile feature to implement MVI architecture and improve module organization
