+++ feature/profile/src/commonMain/kotlin/com/stslex/wizard/feature/profile/ui/mvi/handlers/RepeatLastActionHandler.kt
@@ -1,11 +1,11 @@
-package com.stslex.wizard.feature.profile.mvi
+package com.stslex.wizard.feature.profile.ui.mvi.handlers
 
 
 import com.stslex.wizard.core.ui.mvi.v2.Handler
 import com.stslex.wizard.feature.profile.di.ProfileScope
-import com.stslex.wizard.feature.profile.ui.store.ProfileHandlerStore
-import com.stslex.wizard.feature.profile.ui.store.ProfileScreenState
-import com.stslex.wizard.feature.profile.ui.store.ProfileStore.Action
+import com.stslex.wizard.feature.profile.ui.mvi.store.ProfileHandlerStore
+import com.stslex.wizard.feature.profile.ui.mvi.store.ProfileScreenState
+import com.stslex.wizard.feature.profile.ui.mvi.store.ProfileStore.Action
 import org.koin.core.annotation.Factory
 import org.koin.core.annotation.Scope
 import org.koin.core.annotation.Scoped
