+++ feature/profile/src/commonMain/kotlin/com/stslex/wizard/feature/profile/ui/mvi/store/ProfileStoreImpl.kt
@@ -1,16 +1,16 @@
-package com.stslex.wizard.feature.profile.ui.store
+package com.stslex.wizard.feature.profile.ui.mvi.store
 
 import com.stslex.wizard.core.core.coroutine.AppDispatcher
 import com.stslex.wizard.core.ui.mvi.v2.BaseStore
 import com.stslex.wizard.feature.profile.di.ProfileScope
-import com.stslex.wizard.feature.profile.mvi.ClickersHandler
-import com.stslex.wizard.feature.profile.mvi.InitStorageHandler
-import com.stslex.wizard.feature.profile.mvi.LogoutHandler
-import com.stslex.wizard.feature.profile.mvi.RepeatLastActionHandler
+import com.stslex.wizard.feature.profile.ui.mvi.handlers.ClickersHandler
+import com.stslex.wizard.feature.profile.ui.mvi.handlers.InitStorageHandler
+import com.stslex.wizard.feature.profile.ui.mvi.handlers.LogoutHandler
+import com.stslex.wizard.feature.profile.ui.mvi.handlers.RepeatLastActionHandler
 import com.stslex.wizard.feature.profile.navigation.ProfileComponent
-import com.stslex.wizard.feature.profile.ui.store.ProfileStore.Action
-import com.stslex.wizard.feature.profile.ui.store.ProfileStore.Event
-import com.stslex.wizard.feature.profile.ui.store.ProfileStore.State
+import com.stslex.wizard.feature.profile.ui.mvi.store.ProfileStore.Action
+import com.stslex.wizard.feature.profile.ui.mvi.store.ProfileStore.Event
+import com.stslex.wizard.feature.profile.ui.mvi.store.ProfileStore.State
 import org.koin.android.annotation.KoinViewModel
 import org.koin.core.annotation.Qualifier
 import org.koin.core.annotation.Scope
